Switch jessevdk/go-flags for spf13/cobra
